rename utils folder to lib
